Добавил отзывы к фильму

+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -1,10 +1,8 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
-import ru.yandex.practicum.filmorate.exception.DuplicateException;
 import ru.yandex.practicum.filmorate.exception.InternalServerException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Review;
@@ -23,7 +21,17 @@ public class ReviewService {
         return storage.findById(id).orElseThrow(() -> new NotFoundException("Отзыв не найден"));
     }
 
+    public Review findByFilmAndUserId(Review review) {
+        return storage.findByFilmAndUserId(review).orElseThrow(() -> new NotFoundException("Отзыв не найден"));
+    }
+
     public Review create(Review review) {
+        if (userService.getUserById((long) review.getUserId()) == null) {
+            throw new NotFoundException("Пользователь не найден");
+        }
+        if (filmService.getFilmById((long) review.getFilmId()) == null) {
+            throw new NotFoundException("Фильм не найден");
+        }
         return storage.create(review);
     }
 
@@ -32,7 +40,7 @@ public class ReviewService {
     }
 
     public Review update(Review review) {
-        Review oldReview = findById(review.getId());
+        Review oldReview = findByFilmAndUserId(review);
         if (StringUtils.hasText(review.getContent())) {
             oldReview.setContent(review.getContent());
         }
@@ -61,25 +69,35 @@ public class ReviewService {
             throw new NotFoundException("Пользователь не найден");
         }
         Review review = storage.findById(reviewId).orElseThrow(() -> new NotFoundException("Отзыв не найден"));
-        int rowsAffected = 0;
-        try {
-            rowsAffected = storage.addLikeOrDislike(reviewId, userId, isLike);
-        } catch (DuplicateKeyException e) {
-            throw new DuplicateException("Пользователь уже оценил отзыв");
+        int rowsAffected;
+        if (storage.likeOrDislikeExists(reviewId, userId)) {
+            rowsAffected = storage.updateLikeOrDislike(reviewId, userId, isLike);
+            if (rowsAffected <= 0) {
+                throw new InternalServerException("Не удалось поставить лайк/дизлайк");
+            }
+            storage.changeUseful(review, isLike, 2);
+            return;
         }
+        rowsAffected = storage.addLikeOrDislike(reviewId, userId, isLike);
         if (rowsAffected <= 0) {
             throw new InternalServerException("Не удалось поставить лайк/дизлайк");
         }
-        rowsAffected = storage.changeUseful(review, isLike);
+        rowsAffected = storage.changeUseful(review, isLike, 1);
         if (rowsAffected <= 0) {
             storage.deleteLikeOrDislike(reviewId, userId);
             throw new InternalServerException("Не удалось поставить лайк/дизлайк");
         }
     }
 
-    public void deleteLikeOrDislike(int reviewId, int userId) {
+    public void deleteLikeOrDislike(int reviewId, int userId, boolean isLike) {
+        Review review = storage.findById(reviewId).orElseThrow(() -> new NotFoundException("Отзыв не найден"));
         boolean success = storage.deleteLikeOrDislike(reviewId, userId);
         if (!success) throw new InternalServerException("Не удалось удалить лайк/дизлайк");
+        int rowsAffected = storage.changeUseful(review, !isLike, 1);
+        if (rowsAffected <= 0) {
+            storage.deleteLikeOrDislike(reviewId, userId);
+            throw new InternalServerException("Не удалось удалить лайк/дизлайк");
+        }
     }
 
 }
